requirements.txt
alembic==1.8.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
certifi==2021.10.8
click==8.1.3
cloudpickle==2.1.0
databricks-cli==0.17.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.5
django-extensions==3.1.5
docker==5.0.3
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.2
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
h11==0.13.0
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.6.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mlflow==1.27.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.0
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.2
prompt-toolkit==3.0.30
protobuf==4.21.2
psutil==5.9.1
pure-eval==0.2.2
pycosat==0.6.3
Pygments==2.12.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
querystring-parser==1.2.4
# requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
scipy==1.8.1
selenium==4.3.0
Send2Trash==1.8.0
# six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
stack-data==0.3.0
tabulate==0.8.10
terminado==0.15.0
tinycss2==1.1.1
tornado==6.2
# tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
# urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
waitress==2.1.2
wcwidth==0.2.5
webdriver-manager==3.7.1
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
widgetsnbextension==3.6.1
# win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
wsproto==1.1.0
zipp==3.8.0
